Clean-up

+++ nuxeo-drive-client/nxdrive/engine/processor.py
@@ -854,7 +854,6 @@ class Processor(EngineWorker):
                     log.debug('No local impact of metadata update on'
                               ' document %r.', doc_pair.remote_name)
                 else:
-                    updated_info = None
                     file_or_folder = 'folder' if doc_pair.folderish else 'file'
                     if doc_pair.folderish:
                         self._engine.set_local_folder_lock(doc_pair.local_path)
@@ -866,7 +865,7 @@ class Processor(EngineWorker):
                         old_path = doc_pair.local_path
                         new_path = new_parent_pair.local_path + '/' + moved_name
                         if old_path == new_path:
-                            log.debug('WRONG GUESS FOR MOVE: %r', doc_pair)
+                            log.debug('Wrong guess for move: %r', doc_pair)
                             self._is_remote_move(doc_pair)
                             self._dao.synchronize_state(doc_pair)
 
@@ -911,20 +910,17 @@ class Processor(EngineWorker):
                         self._refresh_local_state(doc_pair, updated_info)
             self._handle_readonly(local_client, doc_pair)
             self._dao.synchronize_state(doc_pair)
-        except (IOError, OSError) as exc:
-            """
-            Delaying local update of remotely modified content %r due to
-            concurrent file access (probably opened by another process)
-            """
-            raise exc
         finally:
-            try:
-                os.remove(self.tmp_file)
-            except (TypeError, OSError):
-                pass
             if doc_pair.folderish:
                 # Release folder lock in any case
                 self._engine.release_folder_lock()
+
+        if not self.tmp_file:
+            return
+        try:
+            os.remove(self.tmp_file)
+        except OSError:
+            pass
 
     def _synchronize_remotely_created(self, doc_pair, local_client, remote_client):
         name = doc_pair.remote_name
@@ -1056,16 +1052,6 @@ class Processor(EngineWorker):
                     local_client.delete_final(doc_pair.local_path)
             self._dao.remove_state(doc_pair)
             self._search_for_dedup(doc_pair)
-        except (IOError, OSError) as e:
-            # Under Windows deletion can be impossible while another
-            # process is accessing the same file (e.g. word processor)
-            # TODO: be more specific as detecting this case:
-            # shall we restrict to the case e.errno == 13 ?
-            log.warning(
-                'Delaying local deletion of remotely deleted item %r due to'
-                ' concurrent file access (probably opened by another process).',
-                doc_pair)
-            raise e
         finally:
             if doc_pair.folderish:
                 self._engine.release_folder_lock()
